return all qos mismatches
